keep track of supplied values through `TimeSlotExtension`
