+++ webview/src/routes/scheduler/ScheduleErrorReport.svelte
@@ -13,7 +13,7 @@
 </script>
 
 {#if reportHasErrors(report)}
-	<div class="card bg-error-400 m-4 grid gap-4 p-4 text-center">
+	<div class="card m-4 grid gap-4 bg-error-400 p-4 text-center">
 		{#if report.target_has_duplicates.length !== 0}
 			<div>
 				<strong>Cannot use targets because of duplicates</strong>
